first try to create base abstract for import-based commands..
